various updates
Status update: After a couple weeks of work, I got an appimage that
was in theory ok, but soon followed a bit letdown: it was unusable.

First of all, it did not show even when it "ran" (according to the
terminal output). This was due to wrong assumption about the Qt
plugin system when in static mode. It took the best part of a day
to figure out how it was intended to work only to discover it was
broken on Linux+CMake.

It took another day to create a system to work around the upstream
bugs (they wont get fixed anytime soon since the last attempt was
reverted due to hard to avoid regressions). This commit add a
simplified version of that system. They "real deal" would be to
use `qmake` to generate the linking order dynamically. This commit
only copy/paste the result instead of calling qmake.

Other than that, the number of system where it runs is still limited
below an acceptable threshold. I am still investigating that one.