Enable the static KF5
I originally wanted to wait before enabling this, but the application
would not start due to some obscure Qt5 corner case where partially
static and dynamic binary on 64bit Linux asserted when it could not
detect it's own QCoreApplication::instance() object between the static
and dynamic parts.

This commit cuts the binary size by 80%, as expect. AppImage size:

 * With Debug 1300mb
 * Stripped (no DCE) 672mb
 * Stripped + DCE (431mb)
 * Static + DCE - plugins (106mb)

Note that ~40mb is still overhead because some frameworks refuse
to be static and force the SHARED flag in their CMakeLists. It's
a bug.

The AppImage will also grow again once the icons and fonts are bundled.